Move snippet extension initialization and rename experiment
